Add generics support
This adds generics support and closes #30.

Adding generics required some care to avoid introducing unnecessary bounds (e.g. we can't just use `derive(Copy, Clone)` on generated structs) as well as to correctly append extra lifetime / const-generic in generated structs, but in the end all seems to work.

In the process I've also optimised the derive a bit by using lazy iterators and references instead of Vecs and cloning in few more places.

I also replaced the custom FieldIdent helper with the built-in syn::Member - this small change is kinda unrelated to the rest, so if you prefer, I can split it out into a separate PR.